refactor: extract admin media endpoints into dedicated router
